requirements.txt

viam-sdk==0.34.0
apriltag
numpy
opencv-python
pillow
typing-extensions
